+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Fixed
 
+- `XIT GIF`: Fix borked gifs
+- `contd-fill-condition-address`: Fix incorrect address selection
+- `focus-buffers-on-click`: Disable this feature in `HQ` to fix relocation input resetting
 - `prun-bugs`: Fix scrollbar gutter in `PROD` taking up space without a scrollbar present
 
 ## 25.3.8